requirements.txt
Django==1.8.3
djangorestframework==3.2.0
mock==1.3.0
